# wino conv realize the input into apply_matrix