eh lazy engine is slow too... had a similar issue in llama. May be fixable in lazy directly. This should stack with linearizer and shapetracker improvements. Post as a draft for reference

master
```
  0 16357.04 ms run, 16356.66 ms python,    0.37 ms CL, 1187.24 loss, 0.002563 LR, 1.45 GB used,     29.02 GFLOPS
  1 4258.19 ms run, 4257.70 ms python,    0.49 ms CL, 1187.59 loss, 0.002577 LR, 5.00 GB used,    109.98 GFLOPS
  2  142.83 ms run,   18.58 ms python,  124.25 ms CL, 2754.43 loss, 0.001741 LR, 5.00 GB used,   3278.85 GFLOPS
  3  142.61 ms run,   11.37 ms python,  131.25 ms CL, 2845.13 loss, 0.000906 LR, 5.00 GB used,   3283.91 GFLOPS
  4  141.55 ms run,    7.28 ms python,  134.27 ms CL, 2122.40 loss, 0.000070 LR, 5.00 GB used,   3308.63 GFLOPS
```
this PR
```
  0 14637.10 ms run, 14636.71 ms python,    0.38 ms CL, 1187.24 loss, 0.002563 LR, 1.45 GB used,     32.00 GFLOPS
  1 3979.29 ms run, 3978.80 ms python,    0.49 ms CL, 1187.59 loss, 0.002577 LR, 5.04 GB used,    117.69 GFLOPS
  2  143.60 ms run,   20.21 ms python,  123.38 ms CL, 2754.43 loss, 0.001741 LR, 5.04 GB used,   3261.34 GFLOPS
  3  144.15 ms run,   11.10 ms python,  133.05 ms CL, 2845.25 loss, 0.000906 LR, 5.04 GB used,   3248.82 GFLOPS
  4  143.38 ms run,    7.51 ms python,  135.87 ms CL, 2122.40 loss, 0.000070 LR, 5.04 GB used,   3266.31 GFLOPS
```

with profiler, filtered to lazy.py, less fanout calls
master
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
 71272/65    0.152    0.000   45.428    0.699 lazy.py:142(realize)
     35/8    0.000    0.000   43.089    5.386 lazy.py:341(_realize_contiguous)
68329/25057    0.153    0.000    4.433    0.000 lazy.py:202(e)
   164607    0.224    0.000    3.161    0.000 lazy.py:99(create_lazybuffer)
68423/24202    0.140    0.000    2.963    0.000 lazy.py:323(_push_movement_ops)
96705/79195    0.106    0.000    2.210    0.000 lazy.py:229(shuffle_and_prune_movement_ops)
    63064    0.054    0.000    1.623    0.000 lazy.py:318(replace_with_movement_ops)
   123439    0.375    0.000    1.513    0.000 lazy.py:116(__init__)
86135/72186    0.079    0.000    0.878    0.000 lazy.py:251(reshape)
    35802    0.051    0.000    0.836    0.000 lazy.py:269(pad)
  5214/68    0.008    0.000    0.720    0.011 lazy.py:280(permute)
14449/2749    0.025    0.000    0.432    0.000 lazy.py:303(shrink)
       81    0.000    0.000    0.168    0.002 lazy.py:244(reduce_op)
       94    0.000    0.000    0.095    0.001 lazy.py:239(_reduce_op)
```
this PR
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
 66836/83    0.137    0.000   42.267    0.509 lazy.py:142(realize)
    35/25    0.000    0.000   39.546    1.582 lazy.py:341(_realize_contiguous)
59353/22705    0.135    0.000    3.617    0.000 lazy.py:202(e)
59447/21850    0.127    0.000    2.917    0.000 lazy.py:323(_push_movement_ops)
    53914    0.306    0.000    2.197    0.000 lazy.py:318(replace_with_movement_ops)
   144285    0.200    0.000    1.909    0.000 lazy.py:99(create_lazybuffer)
85275/70201    0.097    0.000    1.734    0.000 lazy.py:229(shuffle_and_prune_movement_ops)
    31386    0.045    0.000    0.885    0.000 lazy.py:269(pad)
75188/63384    0.071    0.000    0.774    0.000 lazy.py:251(reshape)
   112145    0.267    0.000    0.756    0.000 lazy.py:116(__init__)
11977/2713    0.137    0.000    0.455    0.000 lazy.py:303(shrink)
  5214/68    0.008    0.000    0.441    0.006 lazy.py:280(permute)
       81    0.000    0.000    0.167    0.002 lazy.py:244(reduce_op)
      178    0.005    0.000    0.124    0.001 lazy.py:67(_ast_binaryops)
```