@geohot 2023-09-29
Definitely don't want to do this. A model shouldn't realize until after it's ready to get a real output